Fixed - Remaining Read/Write Issues + TMP Missing
This commit fixes the remaming read/write issues and removes the warnings that were being given for a missing character (a fallback font has also been added for that character).

Annoyingly this only leaves one error at the start of each round which seems unfixable :(

Next, fix the errors that are called once the player fires their bow.