Fixed - Tabbing Out Loses Button Focus
This commit fixes the annoying issues that was present in our GAM130 Demo Day where if the player were to tab-out of the game the controller wouldn't regain button focus. This always forced the player (usually marking lecturers) to have to use the mouse to select a button.

Next, fix the issues that are caused when the player's bow is shot.