Adjusted/Fixed - Arena Size + RandLayout/Slow Menu
This commit firstly adjusts the arena by removing un-needed extras that the player never sees. The main bulk of this commit though is the rectifying of the slow power-up menu which was apparently caused by the RandLayout Script changing layout until a power-up was selected causing infinite resource use.

Next, maybe have a look at the movement.